# backend/openapi/openapi-v0.1.0.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  version: "0.1.0"
  description: |
    Контракты Backend/Frontend для таск-трекера.
    Документ ориентирован на использование в Swagger UI / Redoc и генерацию SDK.

    Базовый префикс всех эндпоинтов: /api/v1
servers:
  - url: http://localhost:8000/api/v1
    description: Основной сервер API
...
security:
  - bearerAuth: []
paths:
...
  /issues/{issueId}/worklogs:
    get:
      operationId: listWorklogs
      tags: [Worklogs]
      summary: Список списаний по задаче
      parameters:
        - $ref: "#/components/parameters/IssueId"
        - name: from
          in: query
          description: Начальная дата периода (включительно).
          schema:
            $ref: "#/components/schemas/Date"
        - name: to
          in: query
          description: Конечная дата периода (включительно).
          schema:
            $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Worklogs + summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedWorklogsWithSummary"
    post:
      operationId: createWorklog
      tags: [Worklogs]
      summary: Добавить списание часов
      parameters:
        - $ref: "#/components/parameters/IssueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorklogRequest"
      responses:
        "201":
          description: Списание создано
          content:
            application/json:
              schema:
                type: object
                required: [worklog]
                properties:
                  worklog:
                    $ref: "#/components/schemas/Worklog"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users/{userId}/worklogs:
    get:
      operationId: listUserWorklogs
      tags: [Worklogs]
      summary: Списания пользователя (для отчётов)
      parameters:
        - $ref: "#/components/parameters/UserId"
        - name: project_id
          in: query
          description: Если указан — фильтр по проекту.
          schema:
            $ref: "#/components/schemas/UUID"
        - name: from
          in: query
          description: Начало периода.
          schema:
            $ref: "#/components/schemas/Date"
        - name: to
          in: query
          description: Конец периода.
          schema:
            $ref: "#/components/schemas/Date"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Списания пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedWorklogs"
  /worklogs/{worklogId}:
    patch:
      operationId: updateWorklog
      tags: [Worklogs]
      summary: Изменить списание
      parameters:
        - $ref: "#/components/parameters/WorklogId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorklogRequest"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                type: object
                required: [worklog]
                properties:
                  worklog:
                    $ref: "#/components/schemas/Worklog"
    delete:
      operationId: deleteWorklog
      tags: [Worklogs]
      summary: Удалить списание
      parameters:
        - $ref: "#/components/parameters/WorklogId"
      responses:
        "204":
          description: Удалено
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token в заголовке Authorization: Bearer <token>"
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице (пагинация).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка (пагинация).
      schema:
        type: integer
        minimum: 0
        default: 0
...
    UserId:
      name: userId
      in: path
      required: true
      description: UUID пользователя.
      schema:
        $ref: "#/components/schemas/UUID"
...
    IssueId:
      name: issueId
      in: path
      required: true
      description: UUID задачи/эпика.
      schema:
        $ref: "#/components/schemas/UUID"
...
    WorklogId:
      name: worklogId
      in: path
      required: true
      description: UUID списания.
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Ошибка запроса / ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    UUID:
      type: string
      format: uuid
      description: UUID v4.
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    Date:
      type: string
      format: date
      description: Дата (YYYY-MM-DD).
      example: "2026-03-02"
...
    Numeric:
      type: number
      format: double
      description: Число (допускаются дробные значения).
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          description: Поле запроса.
        reason:
          type: string
          description: Причина.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Код ошибки.
          example: VALIDATION_ERROR
        message:
          type: string
          description: Текст ошибки.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
          description: ID запроса.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    Worklog:
      type: object
      required: [id, issue_id, user_id, work_date, hours]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        issue_id:
          $ref: "#/components/schemas/UUID"
        user_id:
          $ref: "#/components/schemas/UUID"
        work_date:
          $ref: "#/components/schemas/Date"
        hours:
          $ref: "#/components/schemas/Numeric"
        comment:
          type: string
          nullable: true
    CreateWorklogRequest:
      type: object
      required: [work_date, hours]
      properties:
        work_date:
          $ref: "#/components/schemas/Date"
        hours:
          $ref: "#/components/schemas/Numeric"
          minimum: 0.01
        comment:
          type: string
          nullable: true
    UpdateWorklogRequest:
      type: object
      properties:
        work_date:
          $ref: "#/components/schemas/Date"
        hours:
          $ref: "#/components/schemas/Numeric"
        comment:
          type: string
          nullable: true
...
    WorklogSummary:
      type: object
      required: [planned_hours, logged_hours]
      properties:
        planned_hours:
          $ref: "#/components/schemas/Numeric"
          nullable: true
        logged_hours:
          $ref: "#/components/schemas/Numeric"
...
    PagedWorklogs:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Worklog"
        total:
          type: integer
    PagedWorklogsWithSummary:
      type: object
      required: [items, total, summary]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Worklog"
        total:
          type: integer
        summary:
          $ref: "#/components/schemas/WorklogSummary"
...